Calculate volatility estimate for a compound

Relative volatility value and category is estimated for specified compound using group contribution methods.

Usage

calc_vol(
  input,
  from = c("mol_path"),
  method = c("simpol1"),
  return_fx_groups = FALSE,
  return_calc_steps = FALSE
)

Arguments

input

a path to a .mol file or a character representation such as SMILES or InChI

from

the form of input (currently only a path to a .mol file is implemented)

method

the method for calculating estimated volatility. Currently only the SIMPOL.1 method is implemented---see simpol1() for more details.

return_fx_groups

When TRUE, includes functional group counts in final dataframe.

return_calc_steps

When TRUE, includes intermediate volatility calculation steps in final dataframe. See Details

Value

a tibble with relative volatility index (rvi) and volatility category (category).

Details

\(\textrm{log}_{10}C^\ast\) is used for the calculated relative volatility index (rvi). \(\textrm{log}_{10}C^\ast = \textrm{log}_{10}(PM/RT)\) where \(P\) is the estimated vapor pressure for the compound, \(M\) is molecular mass of the compound, \(R\) is the universal gas constant, and \(T\) is temperature (293.14K or 20ºC). When return_calc_steps = TRUE, the log of estimated vapor pressure, log10_P, and \(\textrm{log}_{10}(M/RT)\), log_alpha, are also returned.

See also

Examples

mol_path <- mol_example("C16181.mol")
calc_vol(mol_path)
#> # A tibble: 1 × 5
#>   mol_path                                          formula name    rvi category
#>   <chr>                                             <chr>   <chr> <dbl> <chr>   
#> 1 /home/runner/work/_temp/Library/volcalc/extdata/… C6H7Cl… NA     6.98 high    

# Return functional group counts from get_fx_groups()
calc_vol(mol_path,  return_fx_groups = TRUE)
#> # A tibble: 1 × 47
#>   mol_path formula name    rvi category carbons carbons_asa rings_aromatic rings
#>   <chr>    <chr>   <chr> <dbl> <chr>      <int>       <int>          <int> <int>
#> 1 /home/r… C6H7Cl… NA     6.98 high           6           0              0     1
#> # ℹ 38 more variables: carbon_dbl_bonds <dbl>, CCCO_aliphatic_ring <int>,
#> #   hydroxyl_groups <int>, aldehydes <int>, ketones <int>, carbox_acids <int>,
#> #   ester <int>, ether <int>, ether_alicyclic <int>, ether_aromatic <int>,
#> #   nitrate <int>, nitro <int>, hydroxyl_aromatic <int>, amine_primary <int>,
#> #   amine_secondary <int>, amine_tertiary <int>, amine_aromatic <int>,
#> #   amides <int>, carbonylperoxynitrate <int>, peroxide <int>,
#> #   hydroperoxide <int>, carbonylperoxyacid <int>, nitrophenol <int>, …

# Return intermediate calculations
calc_vol(mol_path, return_calc_steps = TRUE)
#> # A tibble: 1 × 8
#>   mol_path                  formula name    rvi category  mass log_alpha log10_P
#>   <chr>                     <chr>   <chr> <dbl> <chr>    <dbl>     <dbl>   <dbl>
#> 1 /home/runner/work/_temp/… C6H7Cl… NA     6.98 high      272.      10.1   -3.08