Calculate volatility estimate for a compound

Relative volatility value and category is estimated for specified compound using group contribution methods.

Usage

calc_vol(
  input,
  from = c("mol_path", "smiles"),
  method = c("meredith", "simpol1"),
  environment = c("clean", "polluted", "soil"),
  return_fx_groups = FALSE,
  return_calc_steps = FALSE
)

Arguments

input

a path to a .mol file or a SMILES string.

from

the form of input. Either "mol_path" or "smiles" (default is "mol_path").

method

the method for calculating estimated volatility. See simpol1() for more details.

environment

the environment for calculating relative volatility categories. RVI thresholds for low, moderate, and high volatility are as follows: "clean" (clean atmosphere, default) -2, 0, 2; "polluted" (polluted atmosphere) 0, 2, 4; "soil" 4, 6, 8.

return_fx_groups

When TRUE, includes functional group counts in final dataframe.

return_calc_steps

When TRUE, includes intermediate volatility calculation steps in final dataframe. See Details.

Value

a tibble with relative volatility index (rvi) and volatility category (category).

Details

\(\textrm{log}_{10}C^\ast\) is used for the calculated relative volatility index (rvi). \(\textrm{log}_{10}C^\ast = \textrm{log}_{10}(PM/RT)\) where \(P\) is the estimated vapor pressure for the compound, \(M\) is molecular mass of the compound, \(R\) is the universal gas constant, and \(T\) is temperature (293.14K or 20ºC). When return_calc_steps = TRUE, the log of estimated vapor pressure, log10_P, and \(\textrm{log}_{10}(M/RT)\), log_alpha, are also returned.

See also

Examples

mol_paths <- mol_example()
calc_vol(mol_paths)
#> # A tibble: 5 × 5
#>   mol_path                                          formula name    rvi category
#>   <chr>                                             <chr>   <chr> <dbl> <fct>   
#> 1 /home/runner/work/_temp/Library/volcalc/extdata/… C6H12O6 D-Gl… -2.84 non-vol…
#> 2 /home/runner/work/_temp/Library/volcalc/extdata/… C12H18… (-)-…  1.84 moderate
#> 3 /home/runner/work/_temp/Library/volcalc/extdata/… C6H7Cl… beta…  6.98 high    
#> 4 /home/runner/work/_temp/Library/volcalc/extdata/… C12H22O Geos…  4.16 high    
#> 5 /home/runner/work/_temp/Library/volcalc/extdata/… C5H8    Isop…  8.84 high    

# Return functional group counts from get_fx_groups()
calc_vol(mol_paths,  return_fx_groups = TRUE)
#> # A tibble: 5 × 51
#>   mol_path           formula name    rvi category exact_mass carbons carbons_asa
#>   <chr>              <chr>   <chr> <dbl> <fct>         <dbl>   <int>       <int>
#> 1 /home/runner/work… C6H12O6 D-Gl… -2.84 non-vol…      180.        6           0
#> 2 /home/runner/work… C12H18… (-)-…  1.84 moderate      210.       12           0
#> 3 /home/runner/work… C6H7Cl… beta…  6.98 high          270.        6           0
#> 4 /home/runner/work… C12H22O Geos…  4.16 high          182.       12           0
#> 5 /home/runner/work… C5H8    Isop…  8.84 high           68.1       5           0
#> # ℹ 43 more variables: rings_aromatic <dbl>, rings <dbl>,
#> #   carbon_dbl_bonds <dbl>, CCCO_aliphatic_ring <int>, hydroxyl_groups <int>,
#> #   aldehydes <int>, ketones <int>, carbox_acids <int>, ester <int>,
#> #   ether <int>, ether_alicyclic <int>, ether_aromatic <int>, nitrate <int>,
#> #   nitro <int>, hydroxyl_aromatic <int>, amine_primary <int>,
#> #   amine_secondary <int>, amine_tertiary <int>, amine_aromatic <int>,
#> #   amide_total <int>, amide_primary <int>, amide_secondary <int>, …

# Return intermediate calculations
calc_vol(mol_paths, return_calc_steps = TRUE)
#> # A tibble: 5 × 8
#>   mol_path       formula name    rvi category molecular_weight log_alpha log10_P
#>   <chr>          <chr>   <chr> <dbl> <fct>               <dbl>     <dbl>   <dbl>
#> 1 /home/runner/… C6H12O6 D-Gl… -2.84 non-vol…            180.       9.87  -12.7 
#> 2 /home/runner/… C12H18… (-)-…  1.84 moderate            210.       9.94   -8.10
#> 3 /home/runner/… C6H7Cl… beta…  6.98 high                272.      10.1    -3.08
#> 4 /home/runner/… C12H22O Geos…  4.16 high                182.       9.88   -5.72
#> 5 /home/runner/… C5H8    Isop…  8.84 high                 68.1      9.45   -0.61