SIMPOL.1 method for calculating estimated volatility

Implements the SIMPOL.1 group contribution method for predicting liquid vapor pressure of organic compounds as described in Pankow & Asher (2008). Users will not usually use this function directly, but rather through calc_vol() which uses this as the default (currently only) method.

Usage

simpol1(fx_groups)

Arguments

fx_groups

a data.frame or tibble with counts of functional groups produced by get_fx_groups() (or manually, with the same column names)

Value

a tibble

Details

The output includes a column for log10_P where \(\textrm{log}_{10} P_{\textrm{L},i}^\circ(T) = \sum_k\nu_{k,i}b_k(T)\), or the sum of the counts of functional groups (\(\nu_{k,i}\)) times the coefficients for each functional group (\(b_K(T)\)). Units are in log10 atmospheres.

Note

The method described in Pankow & Asher (2008) allows for calculations of logP at different temperatures. This implementation currently only calculates values at 20ºC.

References

Pankow, J.F., Asher, W.E., 2008. SIMPOL.1: a simple group contribution method for predicting vapor pressures and enthalpies of vaporization of multifunctional organic compounds. Atmos. Chem. Phys. https://doi.org/10.5194/acp-8-2773-2008

See also

Examples

mol_path <- mol_example("C16181.mol")
sdf <- ChemmineR::read.SDFset(mol_path)
fx_groups <- get_fx_groups(sdf)
simpol1(fx_groups)
#> # A tibble: 1 × 46
#>   formula  name   mass carbons carbons_asa rings_aromatic rings carbon_dbl_bonds
#>   <chr>    <chr> <dbl>   <int>       <int>          <int> <int>            <dbl>
#> 1 C6H7Cl5O NA     272.       6           0              0     1                0
#> # ℹ 38 more variables: CCCO_aliphatic_ring <int>, hydroxyl_groups <int>,
#> #   aldehydes <int>, ketones <int>, carbox_acids <int>, ester <int>,
#> #   ether <int>, ether_alicyclic <int>, ether_aromatic <int>, nitrate <int>,
#> #   nitro <int>, hydroxyl_aromatic <int>, amine_primary <int>,
#> #   amine_secondary <int>, amine_tertiary <int>, amine_aromatic <int>,
#> #   amides <int>, carbonylperoxynitrate <int>, peroxide <int>,
#> #   hydroperoxide <int>, carbonylperoxyacid <int>, nitrophenol <int>, …